+++ Zipadoo/Zipadoo/Extention/Extention.swift
@@ -8,10 +8,16 @@
 // View extension을 위해 import
 import SwiftUI
 
-// View에서 키보드 숨기는 함수 extention
+// View의 Extension
 extension View {
     /// View에서 키보드 숨기는 함수
     func hideKeyboard() {
         UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+    }
+    
+    /// 약속 장소 도착시 띄워주는 Alert
+    func arrivalMessageAlert (isPresented: Binding<Bool>, arrival: ArrivalMsgModel) -> some View {
+        return modifier(ArrivalMessagingModifier(isPresented: isPresented, arrival: arrival)
+        )
     }
 }
