+++ Zipadoo/Zipadoo/Views/ContentView.swift
@@ -8,18 +8,19 @@
 import SwiftUI
 
 struct ContentView: View {
+    
     @EnvironmentObject var alertStore: AlertStore
     @StateObject private var promiseViewModel = PromiseViewModel()
-    @StateObject var viewModel = ContentViewModel()
+    @StateObject var viewModel = AuthStore()
     
     var body: some View {
         
         // 로그인이 되어있지 않은 상태면 로그인 뷰로,
-        if viewModel.userSession == nil {
+        if viewModel.userSession?.uid == nil {
             LoginView()
         } else {
             TabView {
-                // 홈
+                // MARK: - 홈 뷰
                 HomeMainView(user: viewModel.currentUser)
                     .tabItem {
                         Image(systemName: "house.fill")
@@ -29,7 +30,7 @@ struct ContentView: View {
                     .environmentObject(alertStore)
                     .environmentObject(promiseViewModel)
                 
-                // 친구리스트
+                // MARK: - 친구리스트 뷰
                 FriendsView()
                     .tabItem {
                         Image(systemName: "person.2.fill")
@@ -37,7 +38,7 @@ struct ContentView: View {
                     }
                     .tag(1)
                 
-                // 마이페이지
+                // MARK: - 마이페이지 뷰
                 MyPageView()
                     .tabItem {
                         Image(systemName: "person.fill")
@@ -45,7 +46,6 @@ struct ContentView: View {
                     }
                     .tag(2)
                     .environmentObject(promiseViewModel)
-                
             }
             .arrivalMessageAlert(isPresented: $alertStore.isPresentedArrival, arrival: alertStore.arrivalMsgAlert)
         }
