+++ Zipadoo/Zipadoo/Views/LaunchScreen.swift
@@ -7,7 +7,9 @@
 
 import SwiftUI
 
+/// 시작화면 LaunchScreen
 struct LaunchScreen: View {
+    
     var body: some View {
         VStack {
             GeometryReader { geometry in
@@ -16,10 +18,10 @@ struct LaunchScreen: View {
                     .scaledToFit()
                     .frame(width: geometry.size.width, height: geometry.size.height * 0.2)
                     .offset(y: geometry.size.height * 0.38)
-            }
-        }
-    }
-}
+            } // GeometryReader
+        } // VStack
+    } // body
+} // struct
 
 #Preview {
     LaunchScreen()
