+++ Zipadoo/Zipadoo/Views/ProfileImageView.swift
@@ -7,35 +7,11 @@
 
 import SwiftUI
 
-enum ProfileImageSize {
-    case xSmall
-    case small
-    case regular
-    case medium
-    case large
-    
-    var dimension: CGFloat {
-        switch self {
-        case .xSmall:
-            return 30
-        case .small:
-            return 60
-        case .regular:
-            return 80
-        case .medium:
-            return 150
-        case .large:
-            return 200
-        }
-    }
-}
-/*
-let dummyImageString: String = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSW47UEfYkQQFKwJfi2sD9SjB4uMYS6zC_RAw&usqp=CAU"
-*/
 struct ProfileImageView: View {
     
     /// 파베작업 후 받아올 유저의 프로필 이미지
     let imageString: String
+    
     let size: ProfileImageSize
     
     var body: some View {
@@ -48,7 +24,7 @@ struct ProfileImageView: View {
                     .aspectRatio(contentMode: .fill)
                     .frame(width: size.dimension, height: size.dimension)
                     .clipShape(Circle())
-            } else { // placeholder
+            } else {
                 Image(systemName: "person.circle")
                     .resizable()
                     .aspectRatio(contentMode: .fill)
@@ -56,9 +32,9 @@ struct ProfileImageView: View {
                     .clipShape(Circle())
                     .foregroundColor(.gray)
             }
-        }
-    }
-}
+        } // AsyncImage
+    } // body
+} // struct
 
 #Preview {
     ProfileImageView(imageString: "", size: .medium)
